[MRG] Major modification of procurement workflow for purchase and sale. Please see nested history for details (../generic-addons/ rev 33)
