+++ purchase_requisition_extended/__init__.py
@@ -1,3 +1,3 @@
 # -*- coding: utf-8 -*-
-from . import model
-from . import wizard
+import model
+import wizard
