+++ purchase_requisition_extended/static/src/js/web_addons.js
@@ -20,7 +20,7 @@ openerp.purchase_requisition_extended = function(instance) {
         },
         close_bids_selection: function () {
             var self = this;
-            new instance.web.Model('purchase.requisition').call("close_callforbids",[self.dataset.context.active_id,self.dataset.context]).then(function(result) {
+            new instance.web.Model(self.dataset.model).call("close_callforbids",[self.dataset.context.active_id,self.dataset.context]).then(function(result) {
                 self.do_action(result,{on_close:function(){self.do_action('history_back')}});
             });
         },
